new api services (permits, employment), move mock data to services + implement react-query hooks for new dataproducts for fetching
