+++ packages/af-shared/src/lib/api/index.ts
@@ -3,7 +3,9 @@ import * as auth from './services/auth';
 import * as codesets from './services/codesets';
 import * as company from './services/company';
 import * as dataspace from './services/dataspace';
+import * as employment from './services/employment';
 import * as jmf from './services/jmf';
+import * as permits from './services/permits';
 import * as profile from './services/profile';
 
 const api = {
@@ -14,6 +16,8 @@ const api = {
   profile,
   jmf,
   dataspace,
+  permits,
+  employment,
 };
 
 export default api;
